Develop (#14)
* add director functional

* add director functional(small corrections)

* add director functional(small corrections)

* add director functional(small corrections -1st rev)

* add director functional(small corrections -2nd rev)

* add director functional(small corrections -3rd rev)

* add director functional(small corrections -4th rev)

* add feed functional (1th rev)

* feature: реализована функциональность «Общие фильмы»

* fix: реализована функциональность «Общие фильмы»

* Add director after 1st team review

* Add director small refactoring

* Add feed (2nd rev)

* Add feed (3rd rev)

* Add-director after resolving conflicts after merging add-common-films to develop

* Add-director small refactoring

* feat: Функциональность «Поиск»

* add reviews

* add reviews fix

* add reviews fix 1

* add reviews fix final

* feat: добавить фильтрацию по жанрам и годам для популярных фильмов

* Add-feed (4th rev)

* Add-feed (5th rev)

* add remove endpoints

* add remove endpoints fix final

* add remove endpoints fix final 2

* Add-feed (6th rev)

* fix/refactor: изменить метод для получения популярных фильмов по жанру и году

* develop without recommendation

* Recommendation service (#11)

* Recommendation service

* Recommendation service

* Recommendation service

* Recommendation service

* Recommendation service

* Recommendation service

* Recommendation service

* Recommendation service fixed

---------


* Add files via upload

* Delete Filmorate.png

* Update README.md

* develop after 1st review

---------

+++ src/main/java/ru/yandex/practicum/filmorate/dao/UserServiceDb.java
@@ -6,13 +6,12 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.EventType;
+import ru.yandex.practicum.filmorate.model.Operation;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.user.UserService;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.Comparator;
-import java.util.HashSet;
 import java.util.Set;
 import java.util.TreeSet;
 
@@ -24,12 +23,14 @@ public class UserServiceDb implements UserService {
 
     private final JdbcTemplate jdbcTemplate;
     private final UserDbStorage userDbStorage;
+    private final EventDBStorage eventDBStorage;
+
     @Override
     public void addToFriends(Integer userId, Integer friendId) {
-
-        if(userDbStorage.getUserById(userId).isEmpty() || userDbStorage.getUserById(friendId).isEmpty()){
-            throw new NotFoundException("Пользователя с id: " + userId +" или с id: " + friendId + " не существует");
+        if (userDbStorage.getUserById(userId).isEmpty() || userDbStorage.getUserById(friendId).isEmpty()) {
+            throw new NotFoundException("Пользователя с id: " + userId + " или с id: " + friendId + " не существует");
         } else {
+
             String sql = "INSERT INTO user_friends (user_id,friends_with,confirmation) " +
                     "VALUES (?,?, NVL2 " +
                     "((SELECT * FROM user_friends WHERE confirmation = TRUE AND user_id = ? AND friends_with = ?)" +
@@ -42,35 +43,45 @@ public class UserServiceDb implements UserService {
                     , userId
 
             );
+            eventDBStorage.addEventToUserFeed(userId, friendId, EventType.FRIEND, Operation.ADD);
         }
     }
 
     @Override
     public void removeFromFriends(Integer userId, Integer friendId) {
-        String sql = "DELETE FROM user_friends WHERE user_id = ? AND friends_with = ?";
+        if (userDbStorage.getUserById(userId).isPresent() && userDbStorage.getUserById(friendId).isPresent()) {
 
-        jdbcTemplate.update(sql
-                ,userId
-                ,friendId
-        );
+            String sql = "DELETE FROM user_friends WHERE user_id = ? AND friends_with = ?";
 
-        jdbcTemplate.update(sql
-                ,friendId
-                ,userId
-        );
+            jdbcTemplate.update(sql
+                    , userId
+                    , friendId
+            );
+
+            jdbcTemplate.update(sql
+                    , friendId
+                    , userId
+            );
+            eventDBStorage.addEventToUserFeed(userId, friendId, EventType.FRIEND, Operation.REMOVE);
+        } else throw new NotFoundException("Пользователя с id: " + userId + " или с id: " + friendId + " не существует");
     }
 
     @Override
     public Set<User> getFriendsOfUser(Integer userId) {
-        String sql = "SELECT * " +
-                "FROM users " +
-                "WHERE id IN (" +
-                "SELECT friends_with " +
-                "FROM user_friends " +
-                "WHERE user_id = ?) " +
-                "ORDER BY id ";
-        TreeSet<User> users = new TreeSet<>(Comparator.comparing(User::getId));
-        users.addAll(jdbcTemplate.query(sql, (rs, rowNum) -> makeUser(rs), userId));
+        TreeSet<User> users;
+        if (userDbStorage.getUserById(userId).isEmpty()) {
+            throw new NotFoundException("Пользователя с id: " + userId + " не существует");
+        } else {
+            String sql = "SELECT * " +
+                    "FROM users " +
+                    "WHERE id IN (" +
+                    "SELECT friends_with " +
+                    "FROM user_friends " +
+                    "WHERE user_id = ?) " +
+                    "ORDER BY id ";
+        users = new TreeSet<>(Comparator.comparing(User::getId));
+        users.addAll(jdbcTemplate.query(sql, (rs, rowNum) -> userDbStorage.makeUser(rs), userId));
+    }
         return users;
     }
 
@@ -88,23 +99,8 @@ public class UserServiceDb implements UserService {
                 "WHERE user_id = ?))";
 
         TreeSet<User> users = new TreeSet<>(Comparator.comparing(User::getId));
-        users.addAll(jdbcTemplate.query(sql, (rs, rowNum) -> makeUser(rs), userId, otherUserId));
+        users.addAll(jdbcTemplate.query(sql, (rs, rowNum) -> userDbStorage.makeUser(rs), userId, otherUserId));
         return users;
     }
-
-    private Set<Long> getFriends(Integer userId) {
-        String sql = "SELECT * FROM user_friends WHERE user_id = ?";
-        return new HashSet<>(jdbcTemplate.query(sql, (rs, rowNum) -> rs.getLong("friends_with"), userId));
-    }
-
-    private User makeUser(ResultSet rs) throws SQLException {
-        User user = new User();
-        user.setFriends(getFriends(rs.getInt("id")));
-        user.setId(rs.getInt("id"));
-        user.setEmail(rs.getString("email"));
-        user.setLogin(rs.getString("login"));
-        user.setName(rs.getString("name"));
-        user.setBirthday(rs.getDate("birthday").toLocalDate());
-        return user;
-    }
 }
+
